% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{cluster_sort_mat}
\alias{cluster_sort_mat}
\title{Cluster-sort the rows of a matrix}
\usage{
cluster_sort_mat(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A matrix
}
\description{
Cluster-sort the rows of a matrix
}
