% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{cluster_sort_order}
\alias{cluster_sort_order}
\title{Cluster order}
\usage{
cluster_sort_order(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A vector of integers,
  the sort order of rows.
}
\description{
Cluster order
}
