% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compile_mean_hi_lo}
\alias{compile_mean_hi_lo}
\title{Get mean and ribbon edges for plotting, for the column "value"}
\usage{
compile_mean_hi_lo(df)
}
\arguments{
\item{df}{A dataframe with columns generation and value}
}
\value{
A dataframe with columns generation, upper, lower & mean_value
}
\description{
Get mean and ribbon edges for plotting, for the column "value"
}
