% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{conflate_and_sum_wt}
\alias{conflate_and_sum_wt}
\title{Merge matrix rows that are identical outside
of the first column; sum the first column when 
rows are merged}
\usage{
conflate_and_sum_wt(m, w)
}
\arguments{
\item{m}{A matrix}

\item{w}{A vector of weights}
}
\value{
A list, containing 
  $matrix and $weight
}
\description{
Merge matrix rows that are identical outside
of the first column; sum the first column when 
rows are merged
}
