% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_arrangements}
\alias{count_arrangements}
\title{Count the number of ways to place p stones in an m x n grid, under the
constraints that no row is empty; no column is empty; only one stone per
cell; and cell (1,1) is filled.}
\usage{
count_arrangements(m, n, p)
}
\arguments{
\item{m}{An integer. The number of grid rows.}

\item{n}{An integer. The number of grid columns.}

\item{p}{An integer. The number of stones}
}
\value{
An integer. The number of solutions.
}
\description{
Count the number of ways to place p stones in an m x n grid, under the
constraints that no row is empty; no column is empty; only one stone per
cell; and cell (1,1) is filled.
}
