% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_arrangements_i1}
\alias{count_arrangements_i1}
\title{Count the number of ways to place p stones in an m x n grid, under the
constraints that no row is empty; no column is empty; only one stone per
cell; cell (1,1) is filled; and cell (i,1) must also be filled, for some i >
1. This indicates the how many times a cell (i,1) is filled, among all the
solutions to the grid-filling constraints described for
`count_arrangements()`.}
\usage{
count_arrangements_i1(m, n, p)
}
\arguments{
\item{m}{An integer. The number of grid rows.}

\item{n}{An integer. The number of grid columns.}

\item{p}{An integer. The number of stones}
}
\value{
An integer. The number of solutions.
}
\description{
Count the number of ways to place p stones in an m x n grid, under the
constraints that no row is empty; no column is empty; only one stone per
cell; cell (1,1) is filled; and cell (i,1) must also be filled, for some i >
1. This indicates the how many times a cell (i,1) is filled, among all the
solutions to the grid-filling constraints described for
`count_arrangements()`.
}
