% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{count_exponents}
\alias{count_exponents}
\title{Count the number of unique values in rows or columns}
\usage{
count_exponents(m)
}
\arguments{
\item{m}{A matrix}

\item{p_type}{A string}
}
\value{
A vector of counts
}
\description{
Count the number of unique values in rows or columns
}
