% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{evolve_mplat}
\alias{evolve_mplat}
\title{Simulate mplat evolution}
\usage{
evolve_mplat(
  x = NULL,
  param_dict = mplat_params(),
  pivot_type = c("morphosite", "lexeme"),
  foci_lex_selector = c("uniform", "zipf"),
  foci_morphosite_selector = c("uniform", "zipf"),
  n_foci = 1,
  pivot_weighting = c("uniform", "zipf", "classwise"),
  pivot_use_proportion = 0,
  evidence_weighting = c("uniform", "zipf", "classwise"),
  evidence_use_proportion = 1,
  evidence_balance = 1/(neg_evidence_coeff + 1),
  neg_evidence_coeff = 0,
  change_classwise = FALSE,
  halt_method = c("n_generations", "quasi_stability"),
  halt_n = 5000,
  n_rep = 1
)
}
\arguments{
\item{x}{Either a list or a matrix. If a matrix, x is interpreted as an mplat
representing the initial state of the inflectional system; if a list, then
a set of parameters to generate that state using initialise_mplat().}

\item{pivot_type}{A string, giving the model type: morphosite pivot (as in
A&M) or lexeme pivot (as in Esher 2015)}

\item{foci_lex_selector}{A string, giving the method for selecting the foci
lexemes, which change.}

\item{foci_morphosite_selector}{A string, giving the method for selecting the
foci morphosites, which change.}

\item{n_foci}{An integer, giving the number of foci to change in a single
generation.}

\item{pivot_weighting}{A string, giving the method for weighting pivots; used
either for sampling them, or weighting them when all are taken into
account.}

\item{pivot_use_proportion}{An numeric, giving the proportion of the
available pivots to use. If set to 0, then just one pivot is used.}

\item{evidence_weighting}{A string, giving the method for weighting evidence
morphosites; used either for sampling them, or weighting them when all are
taken into account.}

\item{evidence_use_proportion}{An numeric, giving the proportion of the
available evidence morphosites to use. If set to 0, just one is used.}

\item{evidence_balance}{A numeric, between 0 and 1, giving the balance of
negative evidence (max when 0) and positive evidence (max when 1).}

\item{neg_evidence_coeff}{A non-negative numeric, an alternative
parameterisation of evidence balance, giving negative evidence strength as
a multiple of positive evidence.}

\item{change_classwise}{A logical, whether to change whole classes at once.}

\item{halt_method}{A string, giving the method for choosing when to halt the
simulation.}

\item{halt_n}{An integer, indicating either the total number of generations
to evolve through or the number of consecutive generations of
quasi-stability before halting.}

\item{n_rep}{An integer, the number of repetitions of the simulation to
perform}
}
\value{
A list, containing: containing: mplat_0; total_generations; a list of
  the model parameters; a list of final mplats; a dataframe of changes (with
  columns generation, foci_lex, foci_morphosite, a_new); a dataframe of
  statistics.
}
\description{
Simulate mplat evolution
}
