% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{generate_arrangements}
\alias{generate_arrangements}
\title{Generate the solutions to placing p stones in an m x n grid, under the
constraints that no row is empty; no column is empty; only one stone per
cell; and cell (1,1) is filled.}
\usage{
generate_arrangements(m, n, p)
}
\arguments{
\item{m}{An integer. The number of grid rows.}

\item{n}{An integer. The number of grid columns.}

\item{p}{An integer. The number of stones}
}
\value{
A list of matrices. The solutions, where 1 = filled cell.
}
\description{
Generate the solutions to placing p stones in an m x n grid, under the
constraints that no row is empty; no column is empty; only one stone per
cell; and cell (1,1) is filled.
}
