% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_row_joint_H}
\alias{get_row_joint_H}
\title{Get pairwise joint entropies of rows}
\usage{
get_row_joint_H(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A matrix of numerics,
  the joint entropies.
}
\description{
Get pairwise joint entropies of rows
}
