% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_weights}
\alias{get_weights}
\title{Uniform and zipf distribution}
\usage{
get_weights(dist_type, n)
}
\arguments{
\item{dist_type}{A string, giving the distribution type}

\item{n}{An integer, the number of data points}
}
\value{
A vector of numerics of length n
}
\description{
Uniform and zipf distribution
}
