% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{initialise_mplat}
\alias{initialise_mplat}
\title{Initialise a parametrically random mplat}
\usage{
initialise_mplat(
  x = 50,
  polymorphies = rep(6, times = 8),
  allele_distribution = c("uniform", "zipfian"),
  class_distribution = c("uniform", "zipfian"),
  col_class_distribution = c("uniform", "zipfian"),
  max_classes = NULL,
  max_col_classes = NULL
)
}
\arguments{
\item{x}{Either an integer or a list. If a list, x is treated a list of
parameters, else as the number of lexemes.}

\item{polymorphies}{A vector of integers: the number of alleles for each
morphosite.}

\item{allele_distribution}{A numeric, giving the exponent of a powerlaw
distribution for skewing the sampling from a morphosite's alleles. A value
of 0 is uniform; a value of 1 is Zipfian.}

\item{class_distribution}{A numeric, giving the exponent of a powerlaw
distribution for skewing the repetitive sampling of lexical classes.}

\item{col_class_distribution}{A numeric, giving the exponent of a powerlaw
distribution for skewing the repetitive sampling of morphosites, i.e.,
having multiple morphosites repeating the same pattern.}

\item{max_classes}{An integer, placing an upper bound on the number of
distinct classes sampled when class_distribution is not zero.}

\item{max_col_classes}{An integer, placing an upper bound on the number of
distinct morphosites sampled when col_class_distribution is not zero.}
}
\value{
A matrix, the mplat.
}
\description{
Initialise a parametrically random mplat
}
