% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{match_rows}
\alias{match_rows}
\title{Which rows in a matrix are identical
to a given vector}
\usage{
match_rows(m, v)
}
\arguments{
\item{m}{A matrix}

\item{v}{A vector, of length ncol(m)}
}
\value{
A vector of row indices
}
\description{
Which rows in a matrix are identical
to a given vector
}
