% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{match_sorted_rows}
\alias{match_sorted_rows}
\title{Which rows in matrix 1 are in matrix 2
Only works for sorted matrices!}
\usage{
match_sorted_rows(m1, m2)
}
\arguments{
\item{m1}{A matrix}

\item{m2}{A martix}
}
\value{
A vector of row indices
}
\description{
Which rows in matrix 1 are in matrix 2
Only works for sorted matrices!
}
