% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisation.R
\name{mplat_params}
\alias{mplat_params}
\title{Make a list of mplat parameters}
\usage{
mplat_params(
  n_lexemes = 50,
  polymorphies = rep(6, times = 8),
  allele_distribution = "uniform",
  class_distribution = "uniform",
  col_class_distribution = "uniform",
  max_classes = NULL,
  max_col_classes = NULL
)
}
\value{
A list, of parameters for `initialise_mplat()`
}
\description{
Make a list of mplat parameters
}
