% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{multicolourise}
\alias{multicolourise}
\title{Add multiple colours}
\usage{
multicolourise(data, axis = c("columns", "rows"))
}
\arguments{
\item{data}{A matrix, giving an mplat or a dataframe giving a long form
evo_df.}

\item{axis}{A string, indicating which axis to apply multicolour to.}
}
\value{
A list, containing the modified mplat and a vector of colours.
}
\description{
Add multiple colours
}
