% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplat_tools.R
\name{pivot_mplat_longer}
\alias{pivot_mplat_longer}
\title{Change mplat to a long format dataframe}
\usage{
pivot_mplat_longer(mplat)
}
\arguments{
\item{mplat}{A matrix}
}
\value{
A dataframe with columns lexeme, morphosite, allele
}
\description{
Change mplat to a long format dataframe
}
