% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_class_contrasts}
\alias{plot_class_contrasts}
\title{Plot class contrast stats}
\usage{
plot_class_contrasts(
  evolution,
  y_limits = c(1, max(evolution$stats$n_classes)),
  y_breaks = c(1, 5, 20, 50, 100),
  step = c("generation", "change")
)
}
\arguments{
\item{evolution}{A list, the output from evolve_mplat.}

\item{y_limits}{A vector of two numerics. The y axis limits.}

\item{y_breaks}{A vector of numerics. The y axis tick mark locations.}

\item{step}{A string, specifying the kind of steps to take through the
evolution: by generation or by change.}
}
\description{
Plot class contrast stats
}
