% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_evolution}
\alias{plot_evolution}
\title{Animate the evolution of a mplat}
\usage{
plot_evolution(
  evolution,
  which_rep = 1,
  step_method = c("generation", "change"),
  skip_steps = if (substr(step_method[1], 1, 1) == "g") {
    
    evolution$total_generations/100
 } else {
     nrow(evolution$change)/100
 },
  telic_order = FALSE,
  multicolour = FALSE,
  multicolour_axis = c("columns", "rows"),
  fps = 10
)
}
\arguments{
\item{evolution}{A list, the output from evolve_mplat.}

\item{which_rep}{An integer, specifying which repetition to plot.}

\item{step_method}{A string, specifying the kind of steps to take through the
evolution: by generation or by change.}

\item{skip_steps}{An integer, specifying how many steps to skip ahead each
time.}

\item{telic_order}{A logical, Whether to order columns and rows so they
cluster at the final generation.}

\item{multicolour}{A logical. Whether to put column/rows in different
colours.}

\item{multicolour_axis}{A string, which axis to apply multicolour to.}

\item{fps}{An integer}
}
\value{
A dataframe of the evolution in long form.
}
\description{
Animate the evolution of a mplat
}
