% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_exp_contrasts}
\alias{plot_exp_contrasts}
\title{Plot exponent contrast stats}
\usage{
plot_exp_contrasts(
  evolution,
  y_limits = c(1, 5),
  y_breaks = c(1, 1.5, 2, 3, 4, 5),
  step = c("generation", "change")
)
}
\arguments{
\item{evolution}{A list, the output from evolve_mplat.}

\item{y_limits}{A vector of two numerics. The y axis limits.}

\item{y_breaks}{A vector of numerics. The y axis tick mark locations.}

\item{step}{A string, specifying the kind of steps to take through the
evolution: by generation or by change.}
}
\description{
Plot exponent contrast stats
}
