% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_H}
\alias{plot_H}
\title{Plot entropy stats}
\usage{
plot_H(
  evolution,
  y_limits = c(0, 3),
  y_breaks = c(0, 0.05, 0.2, 0.5, 1, 3),
  step = c("generation", "change")
)
}
\arguments{
\item{evolution}{A list, the output from evolve_mplat.}

\item{y_limits}{A vector of two numerics. The y axis limits.}

\item{y_breaks}{A vector of numerics. The y axis tick mark locations.}

\item{step}{A string, specifying the kind of steps to take through the
evolution: by generation or by change.}
}
\description{
Plot entropy stats
}
