% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mplat}
\alias{plot_mplat}
\title{Plot a plat}
\usage{
plot_mplat(
  x,
  cluster_lexemes = FALSE,
  cluster_morphosites = FALSE,
  multicolour = FALSE,
  multicolour_axis = c("columns", "rows")
)
}
\arguments{
\item{x}{A matrix or a list. If a list, it is used as parameters for
initialise_mplat(); if a matrix, then it is an mplot.}

\item{cluster_lexemes}{A logical. Whether to sort the lexemes by similarity.}

\item{cluster_morphosites}{A logical. Whether to sort the morphosites by
similarity.}

\item{multicolour}{A logical. Whether to put column/rows in different
colours.}

\item{multicolour_axis}{A string, which axis to apply multicolour to.}
}
\value{
A ggplot raster plot.
}
\description{
Plots one trait, and only the first wordform per cell
}
