% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_stats}
\alias{plot_stats}
\title{Plot stats}
\usage{
plot_stats(
  evolution,
  suppress_top2 = FALSE,
  turnover_limits = c(0, 200),
  turnover_breaks = c(0, 2, 8, 20, 50, 100, 200),
  classes_limits = c(1, 100),
  classes_breaks = c(1, 2, 4, 8, 20, 50, 100)
)
}
\arguments{
\item{evolution}{A list, the output from evolve_mplat.}

\item{suppress_top2}{A logical, whether to suppress the plotting of the top 2
classes}

\item{turnover_limits}{A vector of two numerics. The y axis limits for the
plot of turnover}

\item{turnover_breaks}{A vector of numerics. The y axis tick mark locations
for the plot of turnover}

\item{classes_limits}{A vector of two numerics. The y axis limits for the
plot of classes}

\item{classes_breaks}{A vector of numerics. The y axis tick mark locations
for the plot of classes}
}
\description{
Plot stats
}
