% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_top2classes}
\alias{plot_top2classes}
\title{Plot top two classes stats}
\usage{
plot_top2classes(evolution, step = c("generation", "change"))
}
\arguments{
\item{evolution}{A list, the output from evolve_mplat.}

\item{step}{A string, specifying the kind of steps to take through the
evolution: by generation or by change.}
}
\description{
Plot top two classes stats
}
