% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{power_dist}
\alias{power_dist}
\title{Powerlaw distribution}
\usage{
power_dist(n, exponent)
}
\arguments{
\item{n}{An integer, the number of items}

\item{exponent}{A numeric}
}
\description{
Powerlaw distribution
}
