% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{repeat_row}
\alias{repeat_row}
\title{Make a matrix of a repeated row}
\usage{
repeat_row(row, n)
}
\arguments{
\item{row}{A vector}

\item{n}{An integer}
}
\value{
A matrix
}
\description{
Make a matrix of a repeated row
}
