% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{replicate_evo}
\alias{replicate_evo}
\title{Evaluate an expression n times, returning results in a list.}
\usage{
replicate_evo(n, expr)
}
\arguments{
\item{n}{An integer}

\item{expr}{An expression}
}
\value{
A list
}
\description{
Evaluate an expression n times, returning results in a list.
}
