% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sample_power_dist}
\alias{sample_power_dist}
\title{Powerlaw sample}
\usage{
sample_power_dist(n, size, exponent)
}
\arguments{
\item{n}{An interger, the number of levels to sample from, with replacement}

\item{size}{The sample size}

\item{exponent}{A numeric}
}
\description{
Powerlaw sample
}
