% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{scalar_mult_cols}
\alias{scalar_mult_cols}
\title{Multiply each column by an amount}
\usage{
scalar_mult_cols(m, x)
}
\arguments{
\item{m}{A matrix}

\item{x}{A vector length ncol(m)}
}
\value{
A matrix
}
\description{
Multiply each column by an amount
}
