% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{scalar_mult_rows}
\alias{scalar_mult_rows}
\title{Multiply each row by an amount}
\usage{
scalar_mult_rows(m, x)
}
\arguments{
\item{m}{A matrix}

\item{x}{A vector length nrow(m)}
}
\value{
A matrix
}
\description{
Multiply each row by an amount
}
