% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evolve.R
\name{select_replacement_alleles}
\alias{select_replacement_alleles}
\title{Select replacement alleles according to the generalised substance or identity
model}
\usage{
select_replacement_alleles(
  mplat,
  foci_loc_b,
  focus_loc_e,
  pivot_weighting,
  pivot_use_proportion,
  evidence_weighting,
  evidence_use_proportion,
  evidence_balance
)
}
\arguments{
\item{mplat}{A matrix of integers representing the current state of the
inflectional system, rotated so that pivots are in columns.}

\item{foci_loc_b}{A vector of integers, column indices for mplat giving the
position of the foci.}

\item{focus_loc_e}{An integers, a row index for mplat giving the positions of
the foci.}

\item{pivot_weighting}{A string, giving the method for weighting pivots; used
either for sampling them, or weighting them when all are taken into
account.}

\item{pivot_use_proportion}{An numeric, giving the propotion of the available
pivots to use.}

\item{evidence_weighting}{A string, giving the method for weighting evidence
cells; used either for sampling them, or weighting them when all are taken
into account.}

\item{evidence_use_proportion}{An numeric, giving the propotion of the
available evidence cells to use.}

\item{evidence_balance}{A numeric, between 0 and 1, giving the balance of
negative evidence (max when 0) and positive evidence (max when 1).}
}
\value{
A vector of integers, the replacements for foci
}
\description{
Select replacement alleles according to the generalised substance or identity
model
}
