% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{shuffle_within_cols}
\alias{shuffle_within_cols}
\title{Shuffle within columns}
\usage{
shuffle_within_cols(m, cols = 1:ncol(m))
}
\arguments{
\item{m}{A matrix.}

\item{cols}{A vector of integers,
the columns to shuffle within.}
}
\value{
A matrix, with each.
  column's contents shuffled
}
\description{
Shuffle within columns
}
