% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{shuffle_within_rows}
\alias{shuffle_within_rows}
\title{Shuffle within rows}
\usage{
shuffle_within_rows(m, rows = 1:nrow(m))
}
\arguments{
\item{m}{A matrix.}

\item{rows}{A vector of integers,
the rows to shuffle within.}
}
\value{
A matrix, with each
  row's contents shuffled.
}
\description{
Shuffle within rows
}
