% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{sort_mat}
\alias{sort_mat}
\title{Sort a matrix by all columns, left to right}
\usage{
sort_mat(m)
}
\arguments{
\item{m}{A matrix}
}
\value{
A matrix, same as m but with
  rows sorted.
}
\description{
Sort a matrix by all columns, left to right
}
