% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{square_mat}
\alias{square_mat}
\title{Turn a vector into square matrix by repeating it
either in every row or every column}
\usage{
square_mat(rows, cols)
}
\arguments{
\item{rows}{A vector}

\item{cols}{A vector}
}
\description{
Turn a vector into square matrix by repeating it
either in every row or every column
}
