% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{str2ints}
\alias{str2ints}
\title{Comma delimited string to vector of numerics}
\usage{
str2ints(s)
}
\arguments{
\item{s}{A string}
}
\value{
A vector of numerics
}
\description{
Comma delimited string to vector of numerics
}
