% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{sum_row_mins_in_lower_triangle}
\alias{sum_row_mins_in_lower_triangle}
\title{Sum rows' minimums in lower triangle}
\usage{
sum_row_mins_in_lower_triangle(m)
}
\arguments{
\item{m}{A matrix}
}
\description{
Sum rows' minimums in lower triangle
}
