% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplat_tools.R
\name{unpack_evolution}
\alias{unpack_evolution}
\title{Unpack evolution}
\usage{
unpack_evolution(
  evolution,
  repetition = 1,
  step_method = c("generation", "change"),
  skip_steps = 10
)
}
\arguments{
\item{evolution}{A list, the output from evolve_mplat.}

\item{repetition}{An integer. Which repetition to unpack.}

\item{step_method}{A string, specifying the kind of steps to take through the
evolution: by generation or by change.}

\item{skip_steps}{An integer, specifying how many steps to skip ahead each
time.}
}
\value{
A dataframe of the evolution in long form.
}
\description{
Unpack evolution
}
