% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplat_tools.R
\name{update_long_mplat}
\alias{update_long_mplat}
\title{Update longform mplat by one or more changes}
\usage{
update_long_mplat(mplat, changes)
}
\arguments{
\item{mplat}{A dataframe}

\item{changes}{A dataframe}
}
\value{
A dataframe, the mplat updated by the changes
}
\description{
Update longform mplat by one or more changes
}
