% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_tools.R
\name{update_mat}
\alias{update_mat}
\title{Replace m1 cells with non-na cells
of m2}
\usage{
update_mat(m1, m2)
}
\arguments{
\item{m1}{A matrix}

\item{m2}{A matrix}
}
\value{
A matrix
}
\description{
Replace m1 cells with non-na cells
of m2
}
