% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplat_tools.R
\name{update_mplat}
\alias{update_mplat}
\title{Update mplat by one or more changes}
\usage{
update_mplat(mplat, changes)
}
\arguments{
\item{mplat}{A matrix}

\item{changes}{A dataframe}
}
\value{
A list, the mplat updated by the changes; a vector of changed rows
  and a vector of changed columns
}
\description{
Update mplat by one or more changes
}
