% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{update_row_joint_H}
\alias{update_row_joint_H}
\title{Update pairwise joint entropies of rows}
\usage{
update_row_joint_H(h_mat, m, rows)
}
\arguments{
\item{h_mat}{A matrix of H}

\item{m}{A matrix, whose contents to get the entropy of.}

\item{rows}{A vector of integers, the rows to update.}
}
\value{
A matrix of numerics, the joint entropies.
}
\description{
Update pairwise joint entropies of rows
}
