% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{which_is_sampled_max}
\alias{which_is_sampled_max}
\title{Which is the maximum value, sampling among equal top}
\usage{
which_is_sampled_max(x)
}
\arguments{
\item{x}{A vector of numerics}
}
\value{
A numeric
}
\description{
Which is the maximum value, sampling among equal top
}
